requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Login
Flask-Migrate==4.0.5
Flask-RESTX==1.3.0
Flask-JWT-Extended==4.6.0
Flask-MonitoringDashboard==3.3.2

gunicorn==21.2.0

python-dotenv==1.0.1
pymysql

scikit-learn==1.3.2
joblib==1.3.2
pandas==1.5.3
numpy
matplotlib==3.7.3
wordcloud==1.9.4
nltk
mlflow

prometheus_client

pytest
pytest-flask

flake8
celery[redis]
redis

